new stuff

+++ env_status.go
@@ -20,14 +20,14 @@ type EnvState struct {
 	sync.Mutex
 	EnvStatus string `json:"env_status"`
 
-	EpisodeId int `json:"episode_id"`
+	EpisodeId float32 `json:"episode_id"`
 
 	Fps float32 `json:"fps"`
 }
 
 //Creates new EnvState struct type to store, environment ID, env status,
 //episode ID and global FPS.
-func NewEnvState(EnvId string, EnvStatus string, EpisodeId int, Fps float32) (*EnvState, error) {
+func NewEnvState(EnvId string, EnvStatus string, EpisodeId float32, Fps float32) (*EnvState, error) {
 	var err error //replace with proper function for errors later
 	envStatus := EnvState{
 		EnvId:     EnvId,
@@ -82,7 +82,7 @@ func (e *EnvState) SetEnvStatus(EnvStatus string) bool {
 	return true
 }
 
-func (e *EnvState) SetEpisodeId(EpisodeId int) bool {
+func (e *EnvState) SetEpisodeId(EpisodeId float32) bool {
 	e.Lock()
 	defer e.Unlock()
 	e.EpisodeId = EpisodeId
@@ -110,7 +110,7 @@ func (e *EnvState) GetEnvStatus() string {
 	return e.EnvStatus
 }
 
-func (e *EnvState) GetEpisodeId() int {
+func (e *EnvState) GetEpisodeId() float32 {
 	e.Lock()
 	defer e.Unlock()
 	return e.EpisodeId
